+++ st2common/st2common/models/db/__init__.py
@@ -14,3 +14,35 @@ def setup(db_name=cfg.CONF.database.db_name, db_host=cfg.CONF.database.host,
 
 def teardown():
     disconnect()
+
+
+class MongoDBAccess(object):
+    """
+    Db Object Access class. Provides general implementation which should be
+    specialized for a model type.
+    """
+    def __init__(self, model_kls):
+        self._model_kls = model_kls
+
+    def get_by_name(self, value):
+        for model_object in self._model_kls.objects(name=value):
+            return model_object
+        raise ValueError('{} with name "{}" does not exist.'.format(
+            self._model_kls.__name__, value))
+
+    def get_by_id(self, value):
+        for model_object in self._model_kls.objects(id=value):
+            return model_object
+        raise ValueError('{} with id "{}" does not exist.'.format(
+            self._model_kls.__name__, value))
+
+    def get_all(self):
+        return self._model_kls.objects()
+
+    @staticmethod
+    def add_or_update(model_object):
+        return model_object.save()
+
+    @staticmethod
+    def delete(model_object):
+        model_object.delete()
