+++ st2common/st2common/persistence/stactioncontroller.py
@@ -1,0 +1,10 @@
+from st2common.persistence import Access
+from st2common.models.db.stactioncontroller import staction_access
+
+
+class Staction(Access):
+    IMPL = staction_access
+
+    @classmethod
+    def _get_impl(cls):
+        return cls.IMPL
